Project TODO list added
There has been work done on several fronts at once, lately, which
led to a bit of scattering of ideas and loosing grasp of what
was done and what wasn't.
I therefore hereby establish the TODO list to help keep track
of things that need to wait a bit at the moment.

IMPORTANT NOTE:
this doesn't meen at all that commits that break things are accepted.
Commits must still be consistent; read the notice in the TODO file itself.