proc::logger: Debugging
Fixed log file open (flags and creation attributes).
Added aio_write result (error) resolution.
Note that now, an exception is thrown on an I/O error, which may
(and probably shall) not be a good solution; error in logging I/O
should not cause the process failure...
In the end, we shall probably be more fault tolerant in this matter.

Extended unit test to check logging to STDOUT, STDERR and to file.